+++ app/appgraph/demo.py
@@ -43,16 +43,14 @@ class PurgeProject(graphene.Mutation):
             project_ids_list = [str(project_id)]
 
         output = gclient.execute(gql('''mutation ($projIds:[uuid!]!) {
-            delete_test_source(where:{project_id:{_in:$projIds}}) { affected_rows }
-            m2m_by_conf: delete_test_creator_configuration_m2m (where:{configuration:{project_id:{_in:$projIds}}}) {affected_rows}
-            m2m_by_proj: delete_test_creator_configuration_m2m (where:{project_id:{_in:$projIds}}) {affected_rows}
-            delete_test_creator (where:{test_creator_configuration_m2m:{project_id:{_in:$projIds}}}) {affected_rows}
             delete_configuration_parameter (where:{configuration:{project_id:{_in:$projIds}}}) {affected_rows}
             delete_result_error (where:{execution:{configuration:{project_id:{_in:$projIds}}}}) {affected_rows}
             delete_result_distribution (where:{execution:{configuration:{project_id:{_in:$projIds}}}}) {affected_rows}
             delete_result_aggregate (where:{execution:{configuration:{project_id:{_in:$projIds}}}}) {affected_rows}
             delete_execution (where:{configuration:{project_id:{_in:$projIds}}}) {affected_rows}
             delete_configuration (where:{project_id:{_in:$projIds}}) {affected_rows}
+            delete_test_source(where:{project_id:{_in:$projIds}}) { affected_rows }
+            delete_test_creator (where:{project_id:{_in:$projIds}}) {affected_rows}
             delete_repository (where:{project_id:{_in:$projIds}}) {affected_rows}
             delete_user_project (where:{project_id:{_in:$projIds}}) {affected_rows}
             delete_project(where:{id:{_in:$projIds}}) {affected_rows}
@@ -71,7 +69,6 @@ class DemoProject(graphene.Mutation):
     project_id = graphene.UUID()
 
     def mutate(self, info, name, req_user_id=None):
-        # TODO: add some test_creator inputs when ready, maybe some
 
         UUID = str(uuid.uuid4())
         if not req_user_id:
@@ -82,32 +79,47 @@ class DemoProject(graphene.Mutation):
 
         gclient = hasura_client(current_app.config)
 
-        gclient.execute(gql('''mutation ($id:uuid!, $id2:uuid!, $userId:uuid!, $name:String!, $creatorConfName:String!, $invalidRepoName:String!, $test_creator_data:json!, $request_result:json!, $distribution_result:json!, $timestamp:timestamptz!) {
+        gclient.execute(gql('''mutation (
+                $id:uuid!, $id2:uuid!, $id3:uuid!, $userId:uuid!, $name:String!, $creatorConfName:String!, 
+                $invalidRepoName:String!, $test_creator_data:json!, $request_result:json!, 
+                $distribution_result:json!, $timestamp:timestamptz!
+        ) {
             insert_project (objects:[{id:$id, name:$name}]) {affected_rows}
             insert_user_project (objects:[{id:$id,, project_id:$id, user_id:$userId}]) {affected_rows}
+            
             insert_good_repository: insert_repository (objects:[{id:$id, name:$name, project_id:$id, url:"git@bitbucket.org:acaisoft/load-events.git", type_slug:"load_tests"}]) {affected_rows}
             insert_bad_repository: insert_repository (objects:[{id:$id2, name:$invalidRepoName, project_id:$id, url:"git@bitbucket.org:acaisoft/invalid-url.git", type_slug:"load_tests"}]) {affected_rows}
-            insert_good_conf_repository: insert_configuration (objects:[{id:$id, name:$name, project_id:$id, repository_id:$id, code_source:"repository", type_slug:"load_tests"}]) {affected_rows}
-            insert_bad_conf_repository: insert_configuration (objects:[{name:$invalidRepoName, project_id:$id, repository_id:$id2, code_source:"repository", type_slug:"load_tests"}]) {affected_rows}
+            
+            insert_test_creator(objects:[{ id:$id3, max_wait:200, min_wait:100, data:$test_creator_data, name:$name, project_id:$id  }]) { affected_rows }
+
+            source_3: insert_test_source(objects:[{ id:$id, project_id:$id, source_type:"repository", repository_id:$id }]) { affected_rows }
+            source_2: insert_test_source(objects:[{ id:$id2, project_id:$id, source_type:"repository", repository_id:$id2 }]) { affected_rows }
+            source_1: insert_test_source(objects:[{ id:$id3, project_id:$id, source_type:"creator", test_creator_id:$id3 }]) { affected_rows }
+
+            insert_good_conf_repository: insert_configuration (objects:[{id:$id, name:$name, project_id:$id, test_source_id:$id, type_slug:"load_tests"}]) {affected_rows}
+            insert_bad_conf_repository: insert_configuration (objects:[{id:$id2, name:$invalidRepoName, project_id:$id, test_source_id:$id2, type_slug:"load_tests"}]) {affected_rows}
+            insert_conf_creator: insert_configuration (objects:[{id:$id3, name:$creatorConfName, project_id:$id, test_source_id:$id3, type_slug:"load_tests"}]) {affected_rows}
+            
+            insert_host: insert_configuration_parameter (objects:[{configuration_id:$id, parameter_slug:"load_tests_host", value:"https://att-lwd-go-dev.acaisoft.net/api"}]) {affected_rows}
+            insert_duration: insert_configuration_parameter (objects:[{configuration_id:$id, parameter_slug:"load_tests_duration", value:"15"}]) {affected_rows}
+            insert_conf_bad_host: insert_configuration_parameter (objects:[{configuration_id:$id2, parameter_slug:"load_tests_host", value:"https://att-lwd-go-dev.acaisoft.net/api"}]) {affected_rows}
+            insert_conf_bad_duration: insert_configuration_parameter (objects:[{configuration_id:$id2, parameter_slug:"load_tests_duration", value:"15"}]) {affected_rows}
+            insert_conf_creator_host: insert_configuration_parameter (objects:[{configuration_id:$id3, parameter_slug:"load_tests_host", value:"https://att-lwd-go-dev.acaisoft.net/api"}]) {affected_rows}
+            insert_conf_creator_duration: insert_configuration_parameter (objects:[{configuration_id:$id3, parameter_slug:"load_tests_duration", value:"15"}]) {affected_rows}
+            
             insert_execution (objects:[{id:$id, configuration_id:$id, status:"INIT"}]) {affected_rows}
             insert_result_aggregate (objects:[{execution_id:$id, average_response_time:10, number_of_successes:100, number_of_errors:20, number_of_fails:30, average_response_size:1234}]) {affected_rows}
             insert_result_distribution (objects:[{execution_id:$id, request_result:$request_result, distribution_result:$distribution_result, start:$timestamp, end:$timestamp}]) {affected_rows}
             insert_result_error (objects:[{execution_id:$id, error_type:"AssertionError", name:$name, exception_data:"tralala", number_of_occurrences:120}]) {affected_rows}
-            insert_host: insert_configuration_parameter (objects:[{configuration_id:$id, parameter_slug:"load_tests_host", value:"https://att-lwd-go-dev.acaisoft.net/api"}]) {affected_rows}
-            insert_duration: insert_configuration_parameter (objects:[{configuration_id:$id, parameter_slug:"load_tests_duration", value:"15"}]) {affected_rows}
-            insert_conf_creator: insert_configuration (objects:[{id:$id2, name:$creatorConfName, project_id:$id, code_source:"creator", type_slug:"load_tests"}]) {affected_rows}
-            insert_conf_creator_host: insert_configuration_parameter (objects:[{configuration_id:$id2, parameter_slug:"load_tests_host", value:"https://att-lwd-go-dev.acaisoft.net/api"}]) {affected_rows}
-            insert_conf_creator_duration: insert_configuration_parameter (objects:[{configuration_id:$id2, parameter_slug:"load_tests_duration", value:"15"}]) {affected_rows}
-            insert_test_creator(objects:[{ id:$id, max_wait:200, min_wait:100, data:$test_creator_data }]) { affected_rows }
-            insert_test_creator_configuration_m2m(objects:[{id:$id, configuration_id:$id2, test_creator_id:$id, name:$name, type_slug:"load_tests", project_id:$id }]) { affected_rows }
-
-            source_1: insert_test_source(objects:[{ project_id:$id, source_type:"creator", test_creator_id:$id }]) { affected_rows }
-            source_2: insert_test_source(objects:[{ project_id:$id, source_type:"repository", repository_id:$id }]) { affected_rows }
-            source_3: insert_test_source(objects:[{ project_id:$id, source_type:"repository", repository_id:$id2 }]) { affected_rows }
-
+            
         }'''), variable_values={
-            'id': UUID, 'id2': str(uuid.uuid4()), 'name': name, 'userId': str(req_user_id),
-            'timestamp': datetime.now().astimezone().isoformat(), 'request_result': example_request_result,
+            'id': UUID,
+            'id2': str(uuid.uuid4()),
+            'id3': str(uuid.uuid4()),
+            'name': name,
+            'userId': str(req_user_id),
+            'timestamp': datetime.now().astimezone().isoformat(),
+            'request_result': example_request_result,
             'distribution_result': example_distribution_result,
             'invalidRepoName': f'{name} - invalid repo',
             'creatorConfName': f'{name} - test creator',